requirements.txt
alembic==1.13.3
asyncpg==0.29.0
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.92.0
fastapi-storages[full]==0.3.0
passlib==1.7.4
pillow~=9.4
python-dotenv==1.0.1
python-jose==3.3.0
sqlalchemy==2.0.35
sqladmin[full]==0.19.0
uvicorn==0.30.6
